[kstreams] Migrating aggregator to Quarkus (HotSpot)
