+++ kstreams-live-update/aggregator/src/main/resources/META-INF/resources/index.html
