+++ kstreams-live-update/aggregator/src/main/resources/META-INF/resources/js/chart.js/Chart.js
